+++ libs/framework-core/source/ftrack_framework_core/event/__init__.py
@@ -46,7 +46,7 @@ class EventManager(object):
     '''Manages the events handling.'''
 
     def __repr__(self):
-        return '<EventManager:{}:{}>'.format(self.mode, self.id)
+        return '<EventManager:{}:{}>'.format(self.allow_remote_events, self.id)
 
     def __del__(self):
         self.logger.debug('Closing {}'.format(self))
@@ -71,8 +71,18 @@ class EventManager(object):
         return _connected
 
     @property
-    def mode(self):
-        return self._mode
+    def allow_remote_events(self):
+        return self._allow_remote_events
+
+    @allow_remote_events.setter
+    def allow_remote_events(self, value):
+        if not isinstance(value, bool):
+            self.logger.error('allow_remote_events must be a boolean')
+            return
+        if value:
+            self._connect()
+            self._wait()
+        self._allow_remote_events = value
 
     @property
     def publish(self):
@@ -108,38 +118,36 @@ class EventManager(object):
             # self.logger.debug('Starting new hub thread for {}'.format(self))
             self._event_hub_thread.start()
 
-    def __init__(self, session, mode=constants.event.LOCAL_EVENT_MODE):
+    def close(self):
+        '''Close the event manager and disconnect from the event hub.'''
+        self.session.event_hub.disconnect()
+        self.session.close()
+
+    def __init__(self, session, allow_remote_events=True):
         self.logger = logging.getLogger(
             __name__ + '.' + self.__class__.__name__
         )
         self._event_hub_thread = None
-        self._mode = mode
         self._session = session
-        if mode == constants.event.REMOTE_EVENT_MODE:
-            # TODO: Bring this back when API event hub properly can differentiate between local and remote mode
-            self._connect()
-            self._wait()
+        self.allow_remote_events = allow_remote_events
 
         # Initialize Publish and subscribe classes to be able to provide
         # predefined events.
         self._publish_instance = Publish(self)
         self._subscribe_instance = Subscribe(self)
 
-        # self.logger.debug('Initialising {}'.format(self))
+        self.logger.debug(
+            f'Initialising event manager {self} with remote events {self.allow_remote_events}'
+        )
 
-    def _publish(self, event, callback=None, mode=None):
-        '''Emit *event* and provide *callback* function, in local or remote *mode*.'''
+    def _publish(self, event, callback=None, remote=None):
+        '''Emit *event* and provide *callback* function, in local or remote *remote*.'''
 
-        mode = mode or self.mode
-
-        if mode is constants.event.LOCAL_EVENT_MODE:
+        if not remote:
             result = self.session.event_hub.publish(
                 event,
                 synchronous=True,
             )
-
-            if result:
-                result = result
 
             # Mock async event reply.
             new_event = ftrack_api.event.base.Event(
@@ -154,6 +162,13 @@ class EventManager(object):
             return result
 
         else:
+            if not self.allow_remote_events:
+                self.logger.warning(
+                    'Remote events are not allowed. Please initialize the '
+                    'EventManager with allow_remote_events=True or set the '
+                    'allow_remote_Events property to true'
+                )
+                return
             self.session.event_hub.publish(event, on_reply=callback)
 
     def _subscribe(self, topic, callback):
@@ -170,9 +185,6 @@ class EventManager(object):
         release dangling callback reference in memory'''
         self.session.event_hub.unsubscribe(subscribe_id)
 
-    def available_framework_events(self):
-        pass
-
 
 class Publish(object):
     '''Class with all the events published by the framework'''
@@ -185,7 +197,7 @@ class Publish(object):
         super(Publish, self).__init__()
         self._event_manager = event_manager
 
-    def _publish_event(self, event_topic, data, callback, mode=None):
+    def _publish_event(self, event_topic, data, callback, remote=None):
         '''
         Common method that calls the private publish method from the
         event manager
@@ -194,22 +206,29 @@ class Publish(object):
             topic=event_topic, data=data
         )
         publish_result = self.event_manager._publish(
-            publish_event, callback=callback, mode=mode
+            publish_event, callback=callback, remote=remote
         )
         return publish_result
 
-    def discover_host(self, callback=None):
+    def discover_host(self, callback=None, remote=None):
         '''
         Publish an event with topic
         :const:`~ftrack_framework_core.constants.event.DISCOVER_HOST_TOPIC`
         '''
-
+        # TODO: review this implementation, for now this one can't never be remote as it goes into a loop.
         data = None
         event_topic = constants.event.DISCOVER_HOST_TOPIC
-        return self._publish_event(event_topic, data, callback)
+        return self._publish_event(
+            event_topic, data, callback, remote=False
+        )  # Change this once fixed.
 
     def host_run_tool_config(
-        self, host_id, tool_config_reference, client_options, callback=None
+        self,
+        host_id,
+        tool_config_reference,
+        client_options,
+        callback=None,
+        remote=None,
     ):
         '''
         Publish an event with topic
@@ -221,7 +240,7 @@ class Publish(object):
             'client_options': client_options,
         }
         event_topic = constants.event.HOST_RUN_TOOL_CONFIG_TOPIC
-        return self._publish_event(event_topic, data, callback)
+        return self._publish_event(event_topic, data, callback, remote=remote)
 
     def host_run_ui_hook(
         self,
@@ -231,6 +250,7 @@ class Publish(object):
         client_options,
         payload,
         callback=None,
+        remote=None,
     ):
         '''
         Publish an event with topic
@@ -244,9 +264,11 @@ class Publish(object):
             'payload': payload,
         }
         event_topic = constants.event.HOST_RUN_UI_HOOK_TOPIC
-        return self._publish_event(event_topic, data, callback)
+        return self._publish_event(event_topic, data, callback, remote=remote)
 
-    def host_context_changed(self, host_id, context_id, callback=None):
+    def host_context_changed(
+        self, host_id, context_id, callback=None, remote=None
+    ):
         '''
         Publish an event with topic
         :const:`~ftrack_framework_core.constants.event.HOST_CONTEXT_CHANGED_TOPIC`
@@ -256,9 +278,11 @@ class Publish(object):
             'context_id': context_id,
         }
         event_topic = constants.event.HOST_CONTEXT_CHANGED_TOPIC
-        return self._publish_event(event_topic, data, callback)
+        return self._publish_event(event_topic, data, callback, remote=remote)
 
-    def client_context_changed(self, host_id, context_id, callback=None):
+    def client_context_changed(
+        self, host_id, context_id, callback=None, remote=None
+    ):
         '''
         Publish an event with topic
         :const:`~ftrack_framework_core.constants.event.CLIENT_CONTEXT_CHANGED_TOPIC`
@@ -268,9 +292,11 @@ class Publish(object):
             'context_id': context_id,
         }
         event_topic = constants.event.CLIENT_CONTEXT_CHANGED_TOPIC
-        return self._publish_event(event_topic, data, callback)
+        return self._publish_event(event_topic, data, callback, remote=remote)
 
-    def host_log_item_added(self, host_id, log_item, callback=None):
+    def host_log_item_added(
+        self, host_id, log_item, callback=None, remote=None
+    ):
         '''
         Publish an event with topic
         :const:`~ftrack_framework_core.constants.event.HOST_LOG_ITEM_ADDED_TOPIC`
@@ -281,10 +307,15 @@ class Publish(object):
         }
 
         event_topic = constants.event.HOST_LOG_ITEM_ADDED_TOPIC
-        return self._publish_event(event_topic, data, callback)
+        return self._publish_event(event_topic, data, callback, remote=remote)
 
     def host_run_ui_hook_result(
-        self, host_id, plugin_reference, ui_hook_result, callback=None
+        self,
+        host_id,
+        plugin_reference,
+        ui_hook_result,
+        callback=None,
+        remote=None,
     ):
         '''
         Publish an event with topic
@@ -297,9 +328,11 @@ class Publish(object):
         }
 
         event_topic = constants.event.HOST_UI_HOOK_RESULT_TOPIC
-        return self._publish_event(event_topic, data, callback)
+        return self._publish_event(event_topic, data, callback, remote=remote)
 
-    def client_signal_context_changed(self, client_id, callback=None):
+    def client_signal_context_changed(
+        self, client_id, callback=None, remote=None
+    ):
         '''
         Publish an event with topic
         :const:`~ftrack_framework_core.constants.event.CLIENT_SIGNAL_CONTEXT_CHANGED_TOPIC`
@@ -309,9 +342,11 @@ class Publish(object):
         }
 
         event_topic = constants.event.CLIENT_SIGNAL_CONTEXT_CHANGED_TOPIC
-        return self._publish_event(event_topic, data, callback)
+        return self._publish_event(event_topic, data, callback, remote=remote)
 
-    def client_signal_host_changed(self, client_id, callback=None):
+    def client_signal_host_changed(
+        self, client_id, callback=None, remote=None
+    ):
         '''
         Publish an event with topic
         :const:`~ftrack_framework_core.constants.event.CLIENT_SIGNAL_HOST_CHANGED_TOPIC`
@@ -321,9 +356,11 @@ class Publish(object):
         }
 
         event_topic = constants.event.CLIENT_SIGNAL_HOST_CHANGED_TOPIC
-        return self._publish_event(event_topic, data, callback)
+        return self._publish_event(event_topic, data, callback, remote=remote)
 
-    def client_notify_log_item_added(self, client_id, log_item, callback=None):
+    def client_notify_log_item_added(
+        self, client_id, log_item, callback=None, remote=None
+    ):
         '''
         Publish an event with topic
         :const:`~ftrack_framework_core.constants.event.CLIENT_NOTIFY_LOG_ITEM_ADDED_TOPIC`
@@ -334,10 +371,15 @@ class Publish(object):
         }
 
         event_topic = constants.event.CLIENT_NOTIFY_LOG_ITEM_ADDED_TOPIC
-        return self._publish_event(event_topic, data, callback)
+        return self._publish_event(event_topic, data, callback, remote=remote)
 
     def client_notify_ui_hook_result(
-        self, client_id, plugin_reference, ui_hook_result, callback=None
+        self,
+        client_id,
+        plugin_reference,
+        ui_hook_result,
+        callback=None,
+        remote=None,
     ):
         '''
         Publish an event with topic
@@ -350,14 +392,10 @@ class Publish(object):
         }
 
         event_topic = constants.event.CLIENT_NOTIFY_UI_HOOK_RESULT_TOPIC
-        return self._publish_event(event_topic, data, callback)
+        return self._publish_event(event_topic, data, callback, remote=remote)
 
     def host_verify_plugins(
-        self,
-        host_id,
-        plugin_names,
-        callback=None,
-        mode=constants.event.LOCAL_EVENT_MODE,
+        self, host_id, plugin_names, callback=None, remote=None
     ):
         '''
         Publish an event with topic
@@ -369,7 +407,7 @@ class Publish(object):
         }
 
         event_topic = constants.event.HOST_VERIFY_PLUGINS_TOPIC
-        return self._publish_event(event_topic, data, callback, mode)
+        return self._publish_event(event_topic, data, callback, remote=remote)
 
 
 class Subscribe(object):