+++ projects/framework-premiere/source/ftrack_framework_premiere/__init__.py
@@ -151,11 +151,9 @@ def bootstrap_integration(framework_extensions_path):
         f' {framework_extensions_path}'
     )
 
-    session = ftrack_api.Session(auto_connect_event_hub=False)
+    session = ftrack_api.Session(auto_connect_event_hub=True)
 
-    event_manager = EventManager(
-        session=session, mode=constants.event.LOCAL_EVENT_MODE
-    )
+    event_manager = EventManager(session=session)
 
     registry_instance = registry.Registry()
     registry_instance.scan_extensions(paths=framework_extensions_path)