+++ libs/constants/release_notes.md
@@ -1,5 +1,10 @@
 # ftrack Constants library release Notes
 
+## upcoming
+
+* [feature] framework event; Remove remote/local event constant.
+
+
 ## v2.0.0
 2024-02-12